+++ MK4duo/src/platform/HAL_SAMD/HAL_timers.h
@@ -49,9 +49,7 @@
  *
  * ARDUINO_ARCH_SAM
  */
-
-#ifndef _HAL_TIMERS_DUE_H_
-#define _HAL_TIMERS_DUE_H_
+#pragma once
 
 // --------------------------------------------------------------------------
 // Includes
@@ -79,9 +77,6 @@ typedef struct {
 #define NvicPriorityUart    1
 #define NvicPrioritySystick 15
 
-
-
-
 #define HAL_TIMER_RATE              ((F_CPU)/2) // 24 MHz
 
 #define STEPPER_TIMER               2
@@ -108,18 +103,15 @@ typedef struct {
 #define HAL_ENABLE_ISRs()           ENABLE_STEPPER_INTERRUPT()
 #define HAL_DISABLE_ISRs()          DISABLE_STEPPER_INTERRUPT()
 
-
-#define ISRS_ENABLED()          (!__get_PRIMASK())
-#define ENABLE_ISRS()           __enable_irq()
-#define DISABLE_ISRS()          __disable_irq()
-
+#define ISRS_ENABLED()              (!__get_PRIMASK())
+#define ENABLE_ISRS()               __enable_irq()
+#define DISABLE_ISRS()              __disable_irq()
 
 #if ENABLED(LIN_ADVANCE)
   #define ISR_LA_BASE_CYCLES          64UL
 #else
   #define ISR_LA_BASE_CYCLES          0UL
 #endif
-
 
 // Bezier interpolation adds 40 cycles
 #if ENABLED(BEZIER_JERK_CONTROL)
@@ -143,7 +135,6 @@ typedef struct {
 
 // And each stepper (start + stop pulse) takes in worst case
 #define ISR_STEPPER_CYCLES            16UL
-
 
 // For each stepper, we add its time
 #if HAS_X_STEP
@@ -200,7 +191,6 @@ typedef struct {
 #else
   #define ISR_LA_LOOP_CYCLES  0UL
 #endif
-
 
 // Highly granular delays for step pulses, etc.
 #define DELAY_0_NOP   NOOP
@@ -270,37 +260,29 @@ void HAL_timer_start(const uint8_t timer_num, const uint32_t frequency);
 void HAL_timer_enable_interrupt(const uint8_t timer_num);
 void HAL_timer_disable_interrupt(const uint8_t timer_num);
 
-
 bool HAL_timer_interrupt_is_enabled(const uint8_t timer_num);
 
 FORCE_INLINE static void HAL_timer_set_count(const uint8_t timer_num, hal_timer_t count) {
-    const tTimerConfig * const pConfig = &TimerConfig[timer_num];
-    pConfig->pTimerRegs->COUNT16.CC[0].reg = count;
+  const tTimerConfig * const pConfig = &TimerConfig[timer_num];
+  pConfig->pTimerRegs->COUNT16.CC[0].reg = count;
 }
 
 FORCE_INLINE static hal_timer_t HAL_timer_get_count(const uint8_t timer_num) {
-    const tTimerConfig * const pConfig = &TimerConfig[timer_num];
-    return pConfig->pTimerRegs->COUNT16.CC[0].reg;
+  const tTimerConfig * const pConfig = &TimerConfig[timer_num];
+  return pConfig->pTimerRegs->COUNT16.CC[0].reg;
 }
 
 FORCE_INLINE static void HAL_timer_set_current_count(const uint8_t timer_num, const hal_timer_t count) {
-    
-    const tTimerConfig * const pConfig = &TimerConfig[timer_num];
-    pConfig->pTimerRegs->COUNT16.COUNT.reg = count;
+  const tTimerConfig * const pConfig = &TimerConfig[timer_num];
+  pConfig->pTimerRegs->COUNT16.COUNT.reg = count;
 }
 
 FORCE_INLINE static hal_timer_t HAL_timer_get_current_count(const uint8_t timer_num) {
-    const tTimerConfig * const pConfig = &TimerConfig[timer_num];
-    return pConfig->pTimerRegs->COUNT16.COUNT.reg;
-    
+  const tTimerConfig * const pConfig = &TimerConfig[timer_num];
+  return pConfig->pTimerRegs->COUNT16.COUNT.reg;
 }
+
 FORCE_INLINE static void HAL_timer_isr_prologue(uint8_t timer_num) {
-    const tTimerConfig * const pConfig = &TimerConfig[timer_num];
-    pConfig->pTimerRegs->COUNT16.INTFLAG.bit.MC0 = 1;
+  const tTimerConfig * const pConfig = &TimerConfig[timer_num];
+  pConfig->pTimerRegs->COUNT16.INTFLAG.bit.MC0 = 1;
 }
-
-uint32_t HAL_calc_timer_interval(uint32_t step_rate, uint8_t* loops, const uint8_t scale);
-void HAL_calc_pulse_cycle();
-
-
-#endif /* _HAL_TIMERS_DUE_H_ */